requirements.txt
fastapi
uvicorn
requests
googlesearch-python
trafilatura
